+++ src/api/Shrooms.DataLayer.EntityModels/Models/Lottery/Lottery.cs
@@ -20,7 +20,7 @@ namespace Shrooms.DataLayer.EntityModels.Models.Lottery
 
         public bool IsRefundFailed { get; set; }
 
-        public virtual ImagesCollection Images { get; set; }
+        public virtual ImageCollection Images { get; set; }
 
         public int GiftedTicketLimit { get; set; }
     }
