+++ src/api/Shrooms.Premium/DataTransferObjects/Models/Lotteries/LotteryDetailsDto.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using JetBrains.Annotations;
 using Shrooms.Contracts.Enums;
-using Shrooms.DataLayer.EntityModels.Models.Lottery;
+using Shrooms.DataLayer.EntityModels.Models;
 
 namespace Shrooms.Premium.DataTransferObjects.Models.Lotteries
 {
@@ -19,7 +19,7 @@ namespace Shrooms.Premium.DataTransferObjects.Models.Lotteries
 
         public int EntryFee { get; set; }
 
-        public ImagesCollection Images { get; set; }
+        public ImageCollection Images { get; set; }
 
         public int Participants { get; set; }
