+++ src/api/Shrooms.Premium/DataTransferObjects/Models/Lotteries/LotteryDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Shrooms.Contracts.Enums;
-using Shrooms.DataLayer.EntityModels.Models.Lottery;
+using Shrooms.DataLayer.EntityModels.Models;
 
 namespace Shrooms.Premium.DataTransferObjects.Models.Lotteries
 {
@@ -18,7 +18,7 @@ namespace Shrooms.Premium.DataTransferObjects.Models.Lotteries
 
         public int EntryFee { get; set; }
 
-        public ImagesCollection Images { get; set; }
+        public ImageCollection Images { get; set; }
 
         public int GiftedTicketLimit { get; set; }
     }
