+++ src/api/Shrooms.Premium/Presentation/WebViewModels/Lotteries/LotteryDetailsViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using Shrooms.Contracts.Enums;
-using Shrooms.DataLayer.EntityModels.Models.Lottery;
+using Shrooms.DataLayer.EntityModels.Models;
 using System;
 
 namespace Shrooms.Premium.Presentation.WebViewModels.Lotteries
@@ -18,7 +18,7 @@ namespace Shrooms.Premium.Presentation.WebViewModels.Lotteries
 
         public LotteryStatus Status { get; set; }
 
-        public ImagesCollection Images { get; set; }
+        public ImageCollection Images { get; set; }
 
         public int Participants { get; set; }
